Allow capture of test task custom values to be disabled
