+++ common-custom-user-data-gradle-plugin/src/main/java/com/gradle/CustomBuildScanConfig.java
@@ -1,11 +1,8 @@
 package com.gradle;
 
 import com.gradle.scan.plugin.BuildScanExtension;
-import org.gradle.api.Action;
 import org.gradle.api.Project;
-import org.gradle.api.Task;
 import org.gradle.api.invocation.Gradle;
-import org.gradle.api.tasks.testing.Test;
 
 import java.io.IOException;
 import java.util.Properties;
@@ -23,7 +20,6 @@ final class CustomBuildScanConfig {
         tagCiOrLocal(buildScan);
         addCiMetadata(buildScan);
         addGitMetadata(buildScan);
-        captureTestProperties(buildScan, gradle);
 
         CustomBuildScanConfigExtras.configureBuildScan(buildScan, gradle);
     }
@@ -242,22 +238,6 @@ final class CustomBuildScanConfig {
 
     private static boolean isGitInstalled() {
         return execAndCheckSuccess("git", "--version");
-    }
-
-    private static void captureTestProperties(BuildScanExtension buildScan, Gradle gradle) {
-        gradle.allprojects(p ->
-                p.getTasks().withType(Test.class).configureEach(test ->
-                        test.doFirst("capture configuration for build scans", new Action<Task>() {
-                                    @Override
-                                    public void execute(Task task) {
-                                        buildScan.value(test.getIdentityPath() + "#maxParallelForks", String.valueOf(test.getMaxParallelForks()));
-                                        test.getSystemProperties().forEach((key, val) ->
-                                                buildScan.value(test.getIdentityPath() + "#sysProps-" + key, hashValue(val)));
-                                    }
-                                }
-                        )
-                )
-        );
     }
 
     private static void addCustomValueAndSearchLink(BuildScanExtension buildScan, String label, String value) {
